Load the font catalog off the main actor
Move full font registration and discovery to a utility queue while keeping selected and utility font registration synchronous. Publish catalog results on MainActor.

Share the startup load with early font mutations so registration cannot race imports or deletes, and ignore late results after synchronous completion.
